moved to MSSQL Server, have a look at our README for upgrading strategy

+++ OpenNos.DAL.EF/Entities/ItemInstance.cs
@@ -43,7 +43,7 @@ namespace OpenNos.DAL.EF
 
         public short ItemVNum { get; set; }
 
-        public sbyte Rare { get; set; }
+        public short Rare { get; set; }
 
         [Index("IX_SlotAndType", 2, IsUnique = false, Order = 1)]
         public short Slot { get; set; }
