+++ OpenNos.DAL.EF/Entities/NpcMonster.cs
@@ -62,7 +62,7 @@ namespace OpenNos.DAL.EF
 
         public short DamageMinimum { get; set; }
 
-        public sbyte DarkResistance { get; set; }
+        public short DarkResistance { get; set; }
 
         public short DefenceDodge { get; set; }
 
@@ -78,7 +78,7 @@ namespace OpenNos.DAL.EF
 
         public short ElementRate { get; set; }
 
-        public sbyte FireResistance { get; set; }
+        public short FireResistance { get; set; }
 
         public byte HeroLevel { get; set; }
 
@@ -88,7 +88,7 @@ namespace OpenNos.DAL.EF
 
         public byte Level { get; set; }
 
-        public sbyte LightResistance { get; set; }
+        public short LightResistance { get; set; }
 
         public short MagicDefence { get; set; }
 
@@ -123,7 +123,7 @@ namespace OpenNos.DAL.EF
 
         public short VNumRequired { get; set; }
 
-        public sbyte WaterResistance { get; set; }
+        public short WaterResistance { get; set; }
 
         public int XP { get; set; }
